# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Restful API",
        "description": "This API was concepted during classes of the course 'RESTful API with Laravel: Build a Real API with Laravel' that I took from the Udemy platform.",
        "contact": {
            "email": "nathanaellimacpc@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.11.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Default development server"
        },
        {
            "url": "https://brnathanlima-laravel-api.herokuapp.com",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/sellers/{seller}/transactions": {
            "get": {
                "tags": [
                    "Sellers"
                ],
                "summary": "Get list of seller's transactions",
                "description": "Returns list of seller's transactions",
                "operationId": "getSellerTransactionsList",
                "parameters": [
                    {
                        "name": "seller",
                        "in": "path",
                        "description": "Seller id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "quantity",
                        "in": "query",
                        "description": "List all seller's transaction with the specified quantity of products",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "description": "Transaction property to sort sort the data by",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "How many records to return per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Returns when transaction is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Returns when transaction is not authorized to perform this request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This action is unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Returns when there's some problem with the application. Please report to the development team when getting this response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "development": []
                    },
                    {
                        "production": []
                    }
                ]
            }
        },
...
        "/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get list of transactions",
                "description": "Returns list of transactions",
                "operationId": "getTransactionsList",
                "parameters": [
                    {
                        "name": "quantity",
                        "in": "query",
                        "description": "List all transaction with the specified quantity of products",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "description": "Transaction property to sort sort the data by",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "How many records to return per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Returns when transaction is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Returns when transaction is not authorized to perform this request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This action is unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Returns when there's some problem with the application. Please report to the development team when getting this response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "development": []
                    },
                    {
                        "production": []
                    }
                ]
            }
        },
        "/transactions/{transaction}": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get transaction information",
                "description": "Returns transaction data",
                "operationId": "getTransactionById",
                "parameters": [
                    {
                        "name": "transaction",
                        "in": "path",
                        "description": "Transaction id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Returns when transaction is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Returns when transaction is not authorized to perform this request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This action is unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Returns when there's some problem with the application. Please report to the development team when getting this response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "development": []
                    },
                    {
                        "production": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "development": {
                "type": "oauth2",
                "description": "Laravel  passport oauth2 security for development.",
                "scheme": "http",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://127.0.0.1:8000/oauth/authorize",
                        "tokenUrl": "http://127.0.0.1:8000/oauth/token",
                        "refreshUrl": "http://127.0.0.1:8000/token/refresh",
                        "scopes": {
                            "purchase-product": "Create a new transaction for a specific product.",
                            "manage-products": "Create, read, update and delete products (CRUD).",
                            "manage-account": "Read your account data (except password) and modify it. Cannot delete.",
                            "read-general": "Read General information like purchasing categories, purchased products,\n                                    selling products, selling categories, your transactions (purchases ans sales)."
                        }
                    }
                }
            },
            "production": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security for production.",
                "scheme": "httpa",
                "flows": {
                    "password": {
                        "authorizationUrl": "https://brnathanlima-laravel-api.herokuapp.com/oauth/authorize",
                        "tokenUrl": "https://brnathanlima-laravel-api.herokuapp.com/oauth/token",
                        "refreshUrl": "https://brnathanlima-laravel-api.herokuapp.com/token/refresh",
                        "scopes": {
                            "purchase-product": "Create a new transaction for a specific product.",
                            "manage-products": "Create, read, update and delete products (CRUD).",
                            "manage-account": "Read your account data (except password) and modify it. Cannot delete.",
                            "read-general": "Read General information like purchasing categories, purchased products,\n                                    selling products, selling categories, your transactions (purchases ans sales)."
                        }
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}